feature: id 调整 uuid -> int + 召回拆分开

+++ src/main/java/org/bitmagic/ifeed/domain/entity/Article.java
@@ -1,11 +1,7 @@
 package org.bitmagic.ifeed.domain.entity;
 
 import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
+import lombok.*;
 
 import java.time.Instant;
 import java.util.UUID;
@@ -22,7 +18,7 @@ public class Article {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id", nullable = false, updatable = false)
-    private Integer id;
+    private Long id;
 
     @Column(name = "uid", nullable = false, updatable = false, unique = true)
     private UUID uid;
